Created README
This document is extremely important anyone who is included in the github group/team/organisation named "The Grand Coding" MUST read this document. Without reading this document you WILL find it extremely difficult to live your life. Thank You (even if you do not read this document as you are sacrificing your life for the well being of the other seven billion people on this planet, on behalf of them I thank you for giving us extra resources that you have given us).